python3/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==3.1.1         # via google-auth
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
flask==1.0.3
google-api-core[grpc]==1.11.1  # via google-cloud-core, google-cloud-datastore, google-cloud-storage
google-api-python-client==1.7.9  # via gumo-core
google-auth-httplib2==0.0.3  # via google-api-python-client
google-auth==1.6.3        # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-storage
google-cloud-core==1.0.1  # via google-cloud-datastore, google-cloud-storage
google-cloud-datastore==1.8.0  # via gumo-datastore
google-cloud-storage==1.16.0
google-resumable-media==0.3.2  # via google-cloud-storage
googleapis-common-protos==1.6.0  # via google-api-core
grpcio==1.21.1            # via google-api-core
gumo-core==0.0.30
gumo-datastore==0.0.15
httplib2==0.12.3          # via google-api-python-client, google-auth-httplib2
idna==2.8                 # via requests
injector==0.16.2          # via gumo-core
itsdangerous==1.1.0       # via flask
jinja2==2.10.1            # via flask
markupsafe==1.1.1         # via jinja2
protobuf==3.8.0           # via google-api-core, googleapis-common-protos
pyasn1-modules==0.2.5     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
pytz==2019.1              # via google-api-core
pyyaml==5.1               # via gumo-core
requests==2.22.0
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-api-python-client, google-auth, google-resumable-media, grpcio, protobuf
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.25.3           # via requests
werkzeug==0.15.4          # via flask
